app/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose
passlib==1.7.4
bcrypt==3.2.2
pydantic[email]
pandas
pdfplumber
python-docx
python-dotenv
pydantic
pydantic_settings
python-multipart
razorpay
pillow
pytesseract
opencv-python
redis
reportlab
